Add more parser tests
